Refactor out the common TeensySerial code
